+++ packages/back-end/src/services/datasource.ts
@@ -7,12 +7,7 @@ import Databricks from "../integrations/Databricks";
 import Redshift from "../integrations/Redshift";
 import Snowflake from "../integrations/Snowflake";
 import Postgres from "../integrations/Postgres";
-import {
-  InformationSchema,
-  InformationSchemaInterface,
-  SourceIntegrationInterface,
-  TestQueryRow,
-} from "../types/Integration";
+import { SourceIntegrationInterface, TestQueryRow } from "../types/Integration";
 import BigQuery from "../integrations/BigQuery";
 import ClickHouse from "../integrations/ClickHouse";
 import Mixpanel from "../integrations/Mixpanel";
@@ -24,11 +19,6 @@ import {
 } from "../../types/datasource";
 import Mysql from "../integrations/Mysql";
 import Mssql from "../integrations/Mssql";
-import {
-  createInformationSchema,
-  updateInformationSchemaById,
-} from "../models/InformationSchemaModel";
-import { updateDataSource } from "../models/DataSourceModel";
 
 export function decryptDataSourceParams<T = DataSourceParams>(
   encrypted: string
@@ -111,43 +101,6 @@ export function getSourceIntegrationObject(datasource: DataSourceInterface) {
   return obj;
 }
 
-//TODO: Move this to a informationSchema service file.
-export async function fetchTableData(
-  databaseName: string,
-  tableSchema: string,
-  tableName: string,
-  datasource: DataSourceInterface
-): Promise<{ tableData: null | unknown[]; refreshMS: number }> {
-  const integration = getSourceIntegrationObject(datasource);
-
-  if (!integration.getTableData) {
-    throw new Error("Table data not supported for this data source");
-  }
-
-  const { tableData, refreshMS } = await integration.getTableData(
-    databaseName,
-    tableSchema,
-    tableName
-  );
-
-  return { tableData, refreshMS };
-}
-
-export async function generateInformationSchema(
-  datasource: DataSourceInterface
-): Promise<{
-  informationSchema: InformationSchema[];
-  refreshMS: number;
-}> {
-  const integration = getSourceIntegrationObject(datasource);
-
-  if (!integration.getInformationSchema) {
-    throw new Error("Information schema not supported for this data source");
-  }
-
-  return await integration.getInformationSchema();
-}
-
 export async function testDataSourceConnection(
   datasource: DataSourceInterface
 ) {
@@ -185,63 +138,4 @@ export async function testQuery(
       sql,
     };
   }
-}
-
-export async function initializeDatasourceInformationSchema(
-  datasource: DataSourceInterface,
-  organization: string
-): Promise<void> {
-  // Create an empty informationSchema
-  const emptyInformationSchema = await createInformationSchema(
-    [],
-    organization,
-    datasource.id
-  );
-
-  // Update the datasource with the informationSchemaId
-  await updateDataSource(datasource.id, organization, {
-    settings: {
-      ...datasource.settings,
-      informationSchemaId: emptyInformationSchema.id,
-    },
-  });
-
-  const { informationSchema, refreshMS } = await generateInformationSchema(
-    datasource
-  );
-
-  // Update the empty informationSchema record with the actual informationSchema
-  await updateInformationSchemaById(organization, emptyInformationSchema.id, {
-    ...emptyInformationSchema,
-    databases: informationSchema,
-    status: "COMPLETE",
-    refreshMS,
-  });
-}
-
-export async function updateDatasourceInformationSchema(
-  datasource: DataSourceInterface,
-  organization: string,
-  informationSchema: InformationSchemaInterface
-): Promise<void> {
-  // Reset the informationSchema to remove any errors and change status to "PENDING"
-  await updateInformationSchemaById(organization, informationSchema.id, {
-    ...informationSchema,
-    status: "PENDING",
-    error: undefined,
-  });
-
-  const {
-    informationSchema: updatedInformationSchema,
-    refreshMS,
-  } = await generateInformationSchema(datasource);
-
-  // Update the empty informationSchema record with the actual informationSchema
-  await updateInformationSchemaById(organization, informationSchema.id, {
-    ...informationSchema,
-    databases: updatedInformationSchema,
-    status: "COMPLETE",
-    error: undefined,
-    refreshMS,
-  });
 }
